vertexai/requirements.txt
# Google AI
tensorflow[and-cuda]		# Google Tensorflow with NVIDIA CUDA support
google-cloud-aiplatform		# Google Cloud AI

# Required
matplotlib
mplcursors
numpy
sklearn
pandas